+++ ZenSudoku/ZSFastGameBoard.m
@@ -23,7 +23,7 @@
 	return [self initWithSize:9];
 }
 
-- (id)initWithSize:(int)newSize {
+- (id)initWithSize:(NSInteger)newSize {
 	self = [super init];
 	
 	if (self) {
@@ -39,8 +39,8 @@
 }
 
 - (void)rebuildRowAndColCaches {
-	for (int row = 0; row < size; ++row) {
-		for (int col = 0; col < size; ++col) {
+	for (NSInteger row = 0; row < size; ++row) {
+		for (NSInteger col = 0; col < size; ++col) {
 			// Add the tile to row and col caches.
 			rows[row][col] = &grid[row][col];
 			cols[col][row] = &grid[row][col];
@@ -49,16 +49,16 @@
 }
 
 - (void)rebuildGroupCache {
-	int *totalTilesInEachGroup = malloc(size * sizeof(int));
+	NSInteger *totalTilesInEachGroup = malloc(size * sizeof(NSInteger));
 	
-	for (int i = 0; i < size; ++i) {
+	for (NSInteger i = 0; i < size; ++i) {
 		totalTilesInEachGroup[i] = 0;
 	}
 	
-	for (int row = 0; row < 9; ++row) {
-		for (int col = 0; col < 9; ++col) {
+	for (NSInteger row = 0; row < 9; ++row) {
+		for (NSInteger col = 0; col < 9; ++col) {
 			// Get the groupId on the standard group map.
-			int groupId = grid[row][col].groupId;
+			NSInteger groupId = grid[row][col].groupId;
 			
 			// Add the tile to the group cache.
 			groups[groupId][totalTilesInEachGroup[groupId]] = &grid[row][col];
@@ -72,15 +72,15 @@
 }
 
 - (void)rebuildAllSetsCache {
-	for (int i = 0; i < size; ++i) {
+	for (NSInteger i = 0; i < size; ++i) {
 		allSets[i] = rows[i];
 	}
 	
-	for (int i = 0; i < size; ++i) {
+	for (NSInteger i = 0; i < size; ++i) {
 		allSets[size + i] = cols[i];
 	}
 	
-	for (int i = 0; i < size; ++i) {
+	for (NSInteger i = 0; i < size; ++i) {
 		allSets[(2 * size) + i] = groups[i];
 	}
 }
@@ -93,10 +93,10 @@
 - (void)allocGrid {
 	grid = malloc(size * sizeof(ZSGameTileStub *));
 	
-	for (int i = 0; i < size; ++i) {
+	for (NSInteger i = 0; i < size; ++i) {
 		grid[i] = malloc(size * sizeof(ZSGameTileStub));
 		
-		for (int j = 0; j < size; ++j) {
+		for (NSInteger j = 0; j < size; ++j) {
 			grid[i][j].row = i;
 			grid[i][j].col = j;
 			grid[i][j].groupId = 0;
@@ -106,7 +106,7 @@
 			grid[i][j].totalPencils = 0;
 			grid[i][j].pencils = malloc(size * sizeof(BOOL));
 			
-			for (int p = 0; p < size; ++p) {
+			for (NSInteger p = 0; p < size; ++p) {
 				grid[i][j].pencils[p] = NO;
 			}
 		}
@@ -114,8 +114,8 @@
 }
 
 - (void)freeGrid {
-	for (int i = 0; i < size; ++i) {
-		for (int j = 0; j < size; ++j) {
+	for (NSInteger i = 0; i < size; ++i) {
+		for (NSInteger j = 0; j < size; ++j) {
 			free(grid[i][j].pencils);
 		}
 		
@@ -131,20 +131,20 @@
 	groups = malloc(size * sizeof(ZSGameTileStub **));
 	allSets = malloc(3 * size * sizeof(ZSGameTileStub **));
 	
-	rowContainsAnswer = malloc(size * sizeof(int *));
-	colContainsAnswer = malloc(size * sizeof(int *));
-	groupContainsAnswer = malloc(size * sizeof(int *));
+	rowContainsAnswer = malloc(size * sizeof(NSInteger *));
+	colContainsAnswer = malloc(size * sizeof(NSInteger *));
+	groupContainsAnswer = malloc(size * sizeof(NSInteger *));
 	
-	for (int i = 0; i < size; ++i) {
+	for (NSInteger i = 0; i < size; ++i) {
 		rows[i] = malloc(size * sizeof(ZSGameTileStub *));
 		cols[i] = malloc(size * sizeof(ZSGameTileStub *));
 		groups[i] = malloc(size * sizeof(ZSGameTileStub *));
 		
-		rowContainsAnswer[i] = malloc(size * sizeof(int));
-		colContainsAnswer[i] = malloc(size * sizeof(int));
-		groupContainsAnswer[i] = malloc(size * sizeof(int));
+		rowContainsAnswer[i] = malloc(size * sizeof(NSInteger));
+		colContainsAnswer[i] = malloc(size * sizeof(NSInteger));
+		groupContainsAnswer[i] = malloc(size * sizeof(NSInteger));
 		
-		for (int j = 0; j < size; ++j) {
+		for (NSInteger j = 0; j < size; ++j) {
 			rowContainsAnswer[i][j] = 0;
 			colContainsAnswer[i][j] = 0;
 			groupContainsAnswer[i][j] = 0;
@@ -153,7 +153,7 @@
 }
 
 - (void)freeSetCaches {
-	for (int i = 0; i < size; ++i) {
+	for (NSInteger i = 0; i < size; ++i) {
 		free(rowContainsAnswer[i]);
 		free(colContainsAnswer[i]);
 		free(groupContainsAnswer[i]);
@@ -176,8 +176,8 @@
 #pragma mark - Data Migration
 
 - (void)copyGroupMapFromFastGameBoard:(ZSFastGameBoard *)gameBoard {
-	for (int row = 0; row < size; ++row) {
-		for (int col = 0; col < size; ++col) {
+	for (NSInteger row = 0; row < size; ++row) {
+		for (NSInteger col = 0; col < size; ++col) {
 			grid[row][col].groupId = gameBoard.grid[row][col].groupId;
 		}
 	}
@@ -186,16 +186,16 @@
 }
 
 - (void)copyGuessesFromFastGameBoard:(ZSFastGameBoard *)gameBoard {
-	for (int row = 0; row < size; ++row) {
-		for (int col = 0; col < size; ++col) {
+	for (NSInteger row = 0; row < size; ++row) {
+		for (NSInteger col = 0; col < size; ++col) {
 			[self setGuess:gameBoard.grid[row][col].guess forTileAtRow:row col:col];
 		}
 	}
 }
 
 - (void)copyGroupMapFromGameBoard:(ZSGameBoard *)gameBoard {
-	for (int row = 0; row < size; ++row) {
-		for (int col = 0; col < size; ++col) {
+	for (NSInteger row = 0; row < size; ++row) {
+		for (NSInteger col = 0; col < size; ++col) {
 			grid[row][col].groupId = [gameBoard getTileAtRow:row col:col].groupId;
 		}
 	}
@@ -204,20 +204,20 @@
 }
 
 - (void)copyGuessesFromGameBoard:(ZSGameBoard *)gameBoard {
-	for (int row = 0; row < size; ++row) {
-		for (int col = 0; col < size; ++col) {
+	for (NSInteger row = 0; row < size; ++row) {
+		for (NSInteger col = 0; col < size; ++col) {
 			[self setGuess:[gameBoard getTileAtRow:row col:col].guess forTileAtRow:row col:col];
 		}
 	}
 }
 
 - (void)copyGuessesFromString:(NSString *)guessesString {
-	int currentRow = 0;
-	int currentCol = 0;
+	NSInteger currentRow = 0;
+	NSInteger currentCol = 0;
 	
-	int intEquivalent;
+	NSInteger intEquivalent;
 	
-	for (int i = 0, l = guessesString.length; i < l; ++i) {
+	for (NSInteger i = 0, l = guessesString.length; i < l; ++i) {
 		unichar currentChar = [guessesString characterAtIndex:i];
 		
 		switch (currentChar) {
@@ -235,7 +235,7 @@
 			case '7':
 			case '8':
 			case '9':
-				intEquivalent = (int)currentChar - 48;
+				intEquivalent = (NSInteger)currentChar - 48;
 				[self setGuess:intEquivalent forTileAtRow:currentRow col:currentCol];
 				break;
 				
@@ -255,12 +255,12 @@
 }
 
 - (void)copyGroupMapFromString:(NSString *)groupMapString {
-	int currentRow = 0;
-	int currentCol = 0;
+	NSInteger currentRow = 0;
+	NSInteger currentCol = 0;
 	
-	int intEquivalent;
+	NSInteger intEquivalent;
 	
-	for (int i = 0, l = groupMapString.length; i < l; ++i) {
+	for (NSInteger i = 0, l = groupMapString.length; i < l; ++i) {
 		unichar currentChar = [groupMapString characterAtIndex:i];
 		
 		switch (currentChar) {
@@ -274,7 +274,7 @@
 			case '7':
 			case '8':
 			case '9':
-				intEquivalent = (int)currentChar - 48;
+				intEquivalent = (NSInteger)currentChar - 48;
 				grid[currentRow][currentCol].groupId = intEquivalent;
 				break;
 				
@@ -296,16 +296,16 @@
 }
 
 - (void)copyGuessesToGameBoardAnswers:(ZSGameBoard *)gameBoard {
-	for (int row = 0; row < size; ++row) {
-		for (int col = 0; col < size; ++col) {
+	for (NSInteger row = 0; row < size; ++row) {
+		for (NSInteger col = 0; col < size; ++col) {
 			[gameBoard setAnswer:grid[row][col].guess forTileAtRow:row col:col];
 		}
 	}
 }
 
 - (void)copyGuessesToGameBoardGuesses:(ZSGameBoard *)gameBoard {
-	for (int row = 0; row < size; ++row) {
-		for (int col = 0; col < size; ++col) {
+	for (NSInteger row = 0; row < size; ++row) {
+		for (NSInteger col = 0; col < size; ++col) {
 			[gameBoard setGuess:grid[row][col].guess forTileAtRow:row col:col];
 		}
 	}
@@ -313,8 +313,8 @@
 
 #pragma mark - Setters
 
-- (void)setGuess:(int)guess forTileAtRow:(int)row col:(int)col {
-	int formerGuess = grid[row][col].guess;
+- (void)setGuess:(NSInteger)guess forTileAtRow:(NSInteger)row col:(NSInteger)col {
+	NSInteger formerGuess = grid[row][col].guess;
 	grid[row][col].guess = guess;
 	
 	[self setAllPencils:NO forTileAtRow:row col:col];
@@ -332,19 +332,19 @@
 	}
 }
 
-- (void)clearGuessForTileAtRow:(int)row col:(int)col {
+- (void)clearGuessForTileAtRow:(NSInteger)row col:(NSInteger)col {
 	[self setGuess:0 forTileAtRow:row col:col];
 }
 
 - (void)clearAllGuesses {
-	for (int row = 0; row < size; ++row) {
-		for (int col = 0; col < size; ++col) {
+	for (NSInteger row = 0; row < size; ++row) {
+		for (NSInteger col = 0; col < size; ++col) {
 			[self clearGuessForTileAtRow:row col:col];
 		}
 	}
 }
 
-- (void)setPencil:(BOOL)isSet forPencilNumber:(int)pencilNumber forTileAtRow:(int)row col:(int)col {
+- (void)setPencil:(BOOL)isSet forPencilNumber:(NSInteger)pencilNumber forTileAtRow:(NSInteger)row col:(NSInteger)col {
 	if (!grid[row][col].pencils[pencilNumber - 1] && isSet) {
 		grid[row][col].totalPencils++;
 	} else if (grid[row][col].pencils[pencilNumber - 1] && !isSet) {
@@ -355,33 +355,33 @@
 }
 
 - (void)setAllPencils:(BOOL)isSet {
-	for (int row = 0; row < size; ++row) {
-		for (int col = 0; col < size; ++col) {
+	for (NSInteger row = 0; row < size; ++row) {
+		for (NSInteger col = 0; col < size; ++col) {
 			[self setAllPencils:isSet forTileAtRow:row col:col];
 		}
 	}
 }
 
-- (void)setAllPencils:(BOOL)isSet forTileAtRow:(int)row col:(int)col {
-	for (int i = 1; i <= size; ++i) {
+- (void)setAllPencils:(BOOL)isSet forTileAtRow:(NSInteger)row col:(NSInteger)col {
+	for (NSInteger i = 1; i <= size; ++i) {
 		[self setPencil:isSet forPencilNumber:i forTileAtRow:row col:col];
 	}
 }
 
-- (void)clearInfluencedPencilsForTileAtRow:(int)row col:(int)col {
-	int guess = grid[row][col].guess;
+- (void)clearInfluencedPencilsForTileAtRow:(NSInteger)row col:(NSInteger)col {
+	NSInteger guess = grid[row][col].guess;
 	
 	ZSGameTileStub *tile = &grid[row][col];
 	
-	for (int i = 0; i < size; ++i) {
+	for (NSInteger i = 0; i < size; ++i) {
 		[self setPencil:NO forPencilNumber:guess forTileAtRow:tile->row col:i];
 	}
 	
-	for (int i = 0; i < size; ++i) {
+	for (NSInteger i = 0; i < size; ++i) {
 		[self setPencil:NO forPencilNumber:guess forTileAtRow:i col:tile->col];
 	}
 	
-	for (int i = 0; i < size; ++i) {
+	for (NSInteger i = 0; i < size; ++i) {
 		[self setPencil:NO forPencilNumber:guess forTileAtRow:groups[tile->groupId][i]->row col:groups[tile->groupId][i]->col];
 	}
 }
@@ -389,15 +389,15 @@
 - (void)addAutoPencils {
 	[self setAllPencils:NO];
 	
-	for (int row = 0; row < size; ++row) {
-		for (int col = 0; col < size; ++col) {
+	for (NSInteger row = 0; row < size; ++row) {
+		for (NSInteger col = 0; col < size; ++col) {
 			// Skip the ones with answers.
 			if (grid[row][col].guess) {
 				continue;
 			}
 			
 			// For each possible guess, check to see if it is valid in that tile.
-			for (int guess = 1; guess <= size; ++guess) {
+			for (NSInteger guess = 1; guess <= size; ++guess) {
 				if ([self isGuess:guess validInRow:row col:col]) {
 					[self setPencil:YES forPencilNumber:guess forTileAtRow:row col:col];
 				}
@@ -408,7 +408,7 @@
 
 #pragma mark - Validitiy Checks
 
-- (BOOL)isGuess:(int)guess validInRow:(int)row col:(int)col {
+- (BOOL)isGuess:(NSInteger)guess validInRow:(NSInteger)row col:(NSInteger)col {
 	BOOL validInRow = [self isGuess:guess validInRow:row];
 	BOOL validInCol = [self isGuess:guess validInCol:col];
 	BOOL validInGroup = [self isGuess:guess validInGroup:grid[row][col].groupId];
@@ -416,15 +416,15 @@
 	return validInRow && validInCol && validInGroup;
 }
 
-- (BOOL)isGuess:(int)guess validInRow:(int)row {
+- (BOOL)isGuess:(NSInteger)guess validInRow:(NSInteger)row {
 	return !rowContainsAnswer[row][guess - 1];
 }
 
-- (BOOL)isGuess:(int)guess validInCol:(int)col {
+- (BOOL)isGuess:(NSInteger)guess validInCol:(NSInteger)col {
 	return !colContainsAnswer[col][guess - 1];
 }
 
-- (BOOL)isGuess:(int)guess validInGroup:(int)groupId {
+- (BOOL)isGuess:(NSInteger)guess validInGroup:(NSInteger)groupId {
 	return !groupContainsAnswer[groupId][guess - 1];
 }
 
@@ -432,7 +432,7 @@
 
 - (void)print9x9Grid {
 	NSLog(@" ");
-	for (int row = 0; row < 9; ++row) {
+	for (NSInteger row = 0; row < 9; ++row) {
 		NSLog(@" %i %i %i | %i %i %i | %i %i %i", grid[row][0].guess, grid[row][1].guess, grid[row][2].guess, grid[row][3].guess,
 			  grid[row][4].guess, grid[row][5].guess, grid[row][6].guess, grid[row][7].guess, grid[row][8].guess);
 		
@@ -445,11 +445,11 @@
 
 - (void)print9x9PencilGrid {
 	NSLog(@" ");
-	for (int row = 0; row < 9; ++row) {
+	for (NSInteger row = 0; row < 9; ++row) {
 		NSMutableString *rowString = [NSMutableString string];
 		
-		for (int col = 0; col < 9; ++col) {
-			for (int pencil = 0; pencil < 9; ++pencil) {
+		for (NSInteger col = 0; col < 9; ++col) {
+			for (NSInteger pencil = 0; pencil < 9; ++pencil) {
 				if (grid[row][col].pencils[pencil]) {
 					[rowString appendString:[NSString stringWithFormat:@"%i", (pencil + 1)]];
 				} else {
